Move ConfirmationView to utils
